Forgotten subtraction,now eeepc-900a works.
git-svn-id: https://svn.aros.org/svn/aros/branches/ABI_V0@45170 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ AROS/workbench/devs/monitors/IntelGMA/aros_winsys.c
@@ -57,7 +57,7 @@ extern struct g45staticdata sd;
 #define LOCK_BB          { ObtainSemaphore(&BatchBufferLock); }
 #define UNLOCK_BB        { ReleaseSemaphore(&BatchBufferLock); }
 
-#define BASEADDRESS(p) ((uint32_t)(p) - (intptr_t)sd->Card.Framebuffer)
+#define BASEADDRESS(p) ((uint32_t)(p) - (uint32_t)sd->Card.Framebuffer)
 struct i915_winsys_batchbuffer *batchbuffer_create(struct i915_winsys *iws);
 
 ULONG reserve_status_index()
